implement server heartbeats
In production I've seen instances in which a replicas connection is
half-open; some kind of packet indicating that the replica connection
died was lost somewhere along the way, and the replica trundles merrily
along, unaware that the server has written it off entirely.

Implementing heartbeats should help us avoid that situation by noticing
the situation and restarting the replica.